# swagger-spec.json
{"openapi":"3.0.0","paths":{"/user/findAll":{"get":{"operationId":"UserController_findAllUser","summary":"find all users","parameters":[{"name":"authorization","in":"header","description":"need token","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"success"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource - no access"},"404":{"description":"failed"}},"tags":["User"]}},"/user/findOne":{"get":{"operationId":"UserController_findOneUserById","summary":"find one user","parameters":[{"name":"id","required":true,"in":"query","schema":{"type":"string"}},{"name":"authorization","in":"header","description":"need token","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"user exist"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource - no access"},"404":{"description":"Invalid user id"}},"tags":["User"]}},"/user/delete":{"delete":{"operationId":"UserController_deleteOneUserById","summary":"delete one user","parameters":[{"name":"authorization","in":"header","description":"need token","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"success"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource - no access"},"404":{"description":"user id not exist"}},"tags":["User"]}},"/user/update":{"put":{"operationId":"UserController_update","summary":"update one user","parameters":[{"name":"authorization","in":"header","description":"need token","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"200":{"description":"success"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource - no access"},"404":{"description":"failed"}},"tags":["User"]}},"/user":{"get":{"operationId":"UserController_getUsers","parameters":[{"name":"order","required":false,"in":"query","schema":{"default":"ASC","enum":["ASC","DESC"],"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"take","required":false,"in":"query","schema":{"minimum":1,"maximum":50,"default":10,"type":"number"}},{"name":"q","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Get users list","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageDto"}}}},"401":{"description":"Unauthorized"}},"tags":["User"],"security":[{"bearer":[]}]}},"/user/{id}":{"get":{"operationId":"UserController_getUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get users list","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}}},"401":{"description":"Unauthorized"}},"tags":["User"],"security":[{"bearer":[]}]}},"/auth/signUp":{"post":{"operationId":"AuthController_signUp","summary":"for user sign up","parameters":[],"requestBody":{"required":true,"description":"","content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"201":{"description":"success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}},"400":{"description":"data format error"},"500":{"description":"failed, duplicate email"}},"tags":["Auth"]}},"/auth/signIn":{"post":{"operationId":"AuthController_signIn","summary":"for user sign in","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"201":{"description":"success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInResponseDto"}}}},"400":{"description":"data format error"},"404":{"description":"invalid username or password"}},"tags":["Auth"]}},"/result/findOne":{"get":{"operationId":"ResultController_findOneById","summary":"find one result","parameters":[{"name":"id","required":true,"in":"query","schema":{"type":"string"}},{"name":"authorization","in":"header","description":"need token","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultResponseDto"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource - no access"},"404":{"description":"Invalid result id"}},"tags":["Result"]}},"/result/delete":{"delete":{"operationId":"ResultController_deleteOneResultById","summary":"delete one result","parameters":[{"name":"authorization","in":"header","description":"need token","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"success"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource - no access"},"404":{"description":"result id not exist"}},"tags":["Result"]}},"/result/findAllByEmail":{"get":{"operationId":"ResultController_findAllByEmail","summary":"Get all result of one user","parameters":[{"name":"page","required":true,"in":"query","description":"page","schema":{"type":"string"}},{"name":"amount","required":true,"in":"query","description":"amount","schema":{"type":"string"}},{"name":"authorization","in":"header","description":"need token","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"success","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ResultResponseDto"}}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource - no access"}},"tags":["Result"]}},"/result/update":{"put":{"operationId":"ResultController_update","summary":"update one result","parameters":[{"name":"authorization","in":"header","description":"need token","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultEntity"}}}},"responses":{"201":{"description":"success"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource - no access"},"404":{"description":"failed"}},"tags":["Result"]}},"/result/save":{"post":{"operationId":"ResultController_save","summary":"save one result","parameters":[{"name":"authorization","in":"header","description":"need token","required":true,"schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultEntity"}}}},"responses":{"201":{"description":"success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultResponseDto"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden resource - no access"},"404":{"description":"failed"}},"tags":["Result"]}},"/result/search":{"get":{"operationId":"ResultController_search","summary":"search result from number","parameters":[{"name":"keyword","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"201":{"description":"success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResultResponseDto"}}}},"401":{"description":"Unauthorized"},"404":{"description":"failed"}},"tags":["Result"]}},"/result/test":{"post":{"operationId":"ResultController_test","parameters":[],"responses":{"201":{"description":""}},"tags":["Result"]}}},"info":{"title":"API","description":"","version":"1.0","contact":{}},"tags":[{"name":"User","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SignUpDto":{"type":"object","properties":{"email":{"type":"string","description":"Email"},"fullName":{"type":"string","description":"fullName"},"password":{"type":"string","description":"Password, length: 8-20"}},"required":["email","fullName","password"]},"PageMetaDto":{"type":"object","properties":{"page":{"type":"number"},"take":{"type":"number"},"itemCount":{"type":"number"},"pageCount":{"type":"number"},"hasPreviousPage":{"type":"boolean"},"hasNextPage":{"type":"boolean"}},"required":["page","take","itemCount","pageCount","hasPreviousPage","hasNextPage"]},"PageDto":{"type":"object","properties":{"data":{"type":"array","items":{"type":"array"}},"meta":{"$ref":"#/components/schemas/PageMetaDto"}},"required":["data","meta"]},"UserDto":{"type":"object","properties":{"id":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"fullName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"salt":{"type":"string"},"role":{"type":"string","enum":["USER","ADMIN","BUSINESS"]},"customerId":{"type":"string","description":"For stripe"},"avatar":{"type":"string"},"isActive":{"type":"boolean"}},"required":["id","createdAt","updatedAt"]},"UserResponseDto":{"type":"object","properties":{"id":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"email":{"type":"string","description":"Email"},"fullName":{"type":"string","description":"fullName"},"customerId":{"type":"string","description":"stripe customerId"}},"required":["id","createdAt","updatedAt","email","fullName","customerId"]},"SignInDto":{"type":"object","properties":{"password":{"type":"string","description":"Password, length: 8-20"},"email":{"type":"string","description":"Email"}},"required":["password","email"]},"SignInResponseDto":{"type":"object","properties":{"user":{"description":"User detail","allOf":[{"$ref":"#/components/schemas/UserResponseDto"}]},"token":{"type":"string","description":"JWT token"}},"required":["user","token"]},"ResultResponseDto":{"type":"object","properties":{"email":{"type":"string","description":"Email"},"name":{"type":"string","description":"name"},"fileAddress":{"type":"string","description":"pdf file address"},"receiptAddress":{"type":"string","description":"receipt file address"},"searchNumber":{"type":"string","description":"search number"},"certificate":{"type":"string","description":"certificate number"}},"required":["email","name","fileAddress","receiptAddress","searchNumber","certificate"]},"ResultEntity":{"type":"object","properties":{}}}}}